MAINT: fix expected type hint

+++ turbo_turtle/_cubit_python.py
@@ -655,7 +655,7 @@ def mesh(
     element_type: str,
     output_file: typing.Optional[str] = parsers.mesh_defaults["output_file"],
     part_name: typing.Optional[str] = parsers.mesh_defaults["part_name"],
-    global_seed: typing.Optional[int] = parsers.mesh_defaults["global_seed"],
+    global_seed: typing.Optional[float] = parsers.mesh_defaults["global_seed"],
     edge_seeds: typing.Optional[typing.List] = parsers.mesh_defaults["edge_seeds"]
 ) -> None:
     """Mesh Cubit volumes and sheet bodies by part/volume name
